+++ src/main/java/act/view/excel/ExcelDataFormat.java
@@ -1,0 +1,35 @@
+package act.view.excel;
+
+/*-
+ * #%L
+ * ACT Excel
+ * %%
+ * Copyright (C) 2015 - 2017 ActFramework
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+
+import java.lang.annotation.*;
+
+/**
+ * Mark on a field of a class to specify the Excel data format
+ */
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.FIELD)
+public @interface ExcelDataFormat {
+    /**
+     * Specify the data format of the field
+     */
+    String value();
+}